+++ wp-content/themes/vf-wp-intranet/functions.php
@@ -1,6 +1,5 @@
 <?php 
 
-require_once('vf-wp-intranet-breadcrumbs/index.php');
 require_once('functions/custom-taxonomies.php');
 require_once('functions/cpt-register.php');
